Move files te match global architecture
