+++ app/src/androidTest/java/com/github/se/eduverse/ui/folder/DisplayFolderTest.kt
@@ -7,8 +7,8 @@ import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import com.github.se.eduverse.model.folder.Folder
-import com.github.se.eduverse.model.folder.FolderRepository
-import com.github.se.eduverse.model.folder.FolderViewModel
+import com.github.se.eduverse.repository.FolderRepository
+import com.github.se.eduverse.viewmodel.FolderViewModel
 import com.github.se.eduverse.model.folder.MyFile
 import com.github.se.eduverse.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.github.se.eduverse.ui.navigation.NavigationActions
