+++ app/src/main/java/com/github/se/eduverse/ui/folder/DisplayFolder.kt
@@ -42,7 +42,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.PopupProperties
 import com.github.se.eduverse.model.folder.FilterTypes
-import com.github.se.eduverse.model.folder.FolderViewModel
+import com.github.se.eduverse.viewmodel.FolderViewModel
 import com.github.se.eduverse.ui.navigation.BottomNavigationMenu
 import com.github.se.eduverse.ui.navigation.LIST_TOP_LEVEL_DESTINATION
 import com.github.se.eduverse.ui.navigation.NavigationActions
